combine two same name classes
